@@ -27,40 +27,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.1.2",
-    "@angular/common": "5.1.2",
-    "@angular/compiler": "5.1.2",
-    "@angular/core": "5.1.2",
-    "@angular/forms": "5.1.2",
-    "@angular/platform-browser": "5.1.2",
-    "@angular/platform-browser-dynamic": "5.1.2",
-    "@angular/router": "5.1.2",
-    "@ngrx/effects": "5.2.0",
-    "@ngrx/entity": "5.2.0",
-    "@ngrx/router-store": "5.2.0",
-    "@ngrx/store": "5.2.0",
+    "@angular/animations": "8.2.8",
+    "@angular/common": "8.2.8",
+    "@angular/compiler": "8.2.8",
+    "@angular/core": "8.2.8",
+    "@angular/forms": "8.2.8",
+    "@angular/platform-browser": "8.2.8",
+    "@angular/platform-browser-dynamic": "8.2.8",
+    "@angular/router": "8.2.8",
+    "@ngrx/effects": "8.3.0",
+    "@ngrx/entity": "8.3.0",
+    "@ngrx/router-store": "8.3.0",
+    "@ngrx/store": "8.3.0",
     "classlist.js": "^1.1.20150312",
     "core-js": "2.5.3",
     "date-fns": "1.29.0",
     "font-awesome": "4.7.0",
     "ngx-textarea-autosize": "1.1.1",
     "primeng": "5.2.4",
-    "rxjs": "5.5.6",
+    "rxjs": "6.5.3",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.9.1",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.803.6",
-    "@angular-devkit/core": "0.0.28",
+    "@angular-devkit/core": "8.3.0",
     "@angular/cli": "8.3.6",
-    "@angular/compiler-cli": "5.1.2",
-    "@angular/language-service": "5.1.2",
-    "@ngrx/store-devtools": "5.1.0",
+    "@angular/compiler-cli": "8.2.8",
+    "@angular/language-service": "8.2.8",
+    "@ngrx/store-devtools": "8.3.0",
     "@types/jasmine": "2.5.53",
     "@types/jasminewd2": "2.0.2",
-    "@types/node": "6.0.60",
+    "@types/node": "12.7.8",
     "angular2-template-loader": "0.6.2",
-    "codelyzer": "^5.0.1",
+    "codelyzer": "^5.1.2",
     "husky": "0.14.3",
     "jasmine-core": "2.6.2",
     "jasmine-spec-reporter": "4.1.0",
@@ -72,14 +73,14 @@
     "karma-jasmine-html-reporter": "0.2.2",
     "lint-staged": "6.0.0",
     "ng-apimock": "1.4.2",
-    "ngrx-store-freeze": "0.2.1",
+    "ngrx-store-freeze": "0.2.4",
     "npm-run-all": "4.1.2",
     "prettier": "1.9.1",
     "ts-node": "4.0.2",
     "tslint": "5.8.0",
     "tslint-config-airbnb": "5.4.2",
     "tslint-config-prettier": "1.6.0",
-    "typescript": "2.5.3",
+    "typescript": "3.5.3",
     "wallaby-webpack": "3.9.3"
   },
   "lint-staged": {
